[#42] Move Image and Color to model package
