+++ imageprocessor/imageprocessor-plugin/src/main/java/pl/edu/uj/kimage/plugin/ImageLoaded.java
@@ -1,5 +1,7 @@
 package pl.edu.uj.kimage.plugin;
 
+import pl.edu.uj.kimage.plugin.model.Image;
+
 public class ImageLoaded extends StepResultEvent {
     private final Image loadedImage;
