+++ imageprocessor/imageprocessor-plugin/src/main/java/pl/edu/uj/kimage/plugin/model/Image.java
@@ -1,4 +1,4 @@
-package pl.edu.uj.kimage.plugin;
+package pl.edu.uj.kimage.plugin.model;
 
 public final class Image {
     private final Color[] data;
